# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: Armada Backend API (MVP stub)
  version: 0.1.0
servers:
  - url: http://localhost:4500
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    Player:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    OwnedCampaignProgress:
      type: object
      required: [progress]
      properties:
        progress:
          type: array
          description: Persisted records for the authenticated player, sorted by mission code. An absent record means no saved progress. Client-claimed bestScore is deliberately excluded.
          items:
            type: object
            required: [missionCode, status, lastResult, verifiedResult, verifiedStars]
            properties:
              missionCode:
                type: string
              status:
                type: string
                enum: [NOT_STARTED, IN_PROGRESS, COMPLETED, FAILED]
              lastResult:
                type: object
                nullable: true
                additionalProperties: true
                description: Persisted client result metadata, not an authoritative star rating. Completion status is server-validated separately.
              verifiedResult:
                $ref: '#/components/schemas/VerifiedMissionResult'
              verifiedStars:
                type: integer
                nullable: true
                minimum: 1
                maximum: 3
                description: Server-replayed best star count. Null for legacy or unverified records; never infer stars from lastResult.
    VerifiedMissionResult:
      type: object
      nullable: true
      description: Server-owned best winning replay metadata. Updated atomically with verifiedStars only when the score improves; ties retain the earlier result.
      required: [schemaVersion, result, seed, turnCount, bonusObjectives]
      properties:
        schemaVersion:
          type: integer
          enum: [1]
        result:
          type: string
          enum: [win]
        seed:
          type: integer
          minimum: 0
        turnCount:
          type: integer
          minimum: 1
        bonusObjectives:
          type: object
          additionalProperties:
            type: boolean
    CreatePlayerRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 3
          maxLength: 32
        region:
          type: string
          maxLength: 10
        externalId:
          type: string
...
paths:
...
  /players/me/progress:
    get:
      summary: Read the authenticated player's persisted campaign progress
      description: No player selector or query parameters are accepted. Responses are no-store.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Owned persisted mission records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnedCampaignProgress'
        '400':
          description: Query parameters are not accepted
        '401':
          description: Unauthorized
  /players:
    post:
      summary: Create player
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlayerRequest'
      responses:
        '201':
          description: Player created
          content:
            application/json:
              schema:
                type: object
                properties:
                  player:
                    $ref: '#/components/schemas/Player'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /players/{id}:
    get:
      summary: Fetch player
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Player found
          content:
            application/json:
              schema:
                type: object
                properties:
                  player:
                    $ref: '#/components/schemas/Player'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (ownership)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...